Broke out StringValue implementations
