+++ community/cypher/interpreted-runtime/src/main/scala/org/neo4j/cypher/internal/runtime/interpreted/commands/expressions/Add.scala
@@ -25,8 +25,7 @@ import org.neo4j.cypher.internal.runtime.interpreted.pipes.QueryState
 import org.neo4j.cypher.internal.util.v3_4.CypherTypeException
 import org.neo4j.cypher.internal.util.v3_4.symbols._
 import org.neo4j.values._
-import org.neo4j.values.storable.StringValue.UTF8StringValue
-import org.neo4j.values.storable._
+import org.neo4j.values.storable.{UTF8StringValue, _}
 import org.neo4j.values.utils.UTF8Utils
 import org.neo4j.values.virtual.VirtualValues
