+++ community/values/src/main/java/org/neo4j/values/storable/Values.java
@@ -23,8 +23,6 @@ import java.lang.reflect.Array;
 import java.util.Arrays;
 import java.util.Comparator;
 
-import org.neo4j.values.storable.StringValue.UTF8StringValue;
-
 import static java.lang.String.format;
 
 /**
@@ -121,7 +119,7 @@ public final class Values
 
     public static TextValue stringValue( String value )
     {
-        return new StringValue.Direct( value );
+        return new StringWrappingStringValue( value );
     }
 
     public static Value stringOrNoValue( String value )
@@ -132,7 +130,7 @@ public final class Values
         }
         else
         {
-            return new StringValue.Direct( value );
+            return new StringWrappingStringValue( value );
         }
     }
